Progress on --check-cfg
https://github.com/rust-lang/rfcs/blob/master/text/3013-conditional-compilation-checking.md

The basics work. You can use --check-cfg to specify the set of value
names and values for #[cfg] conditions.

The main remaining work to do is to teach it about the "well-known"
conditions, like "windows", "linux", "target_os", etc. That, and
making sure the diagnostics are good, and writing test cases.