Add stack_protector target option, enabled for various targets
Merged PR 26243: Enable stack protection for msvc targets

Enable stack protection for msvc targets

To comply with the SDL, stack protection should be enabled. Binskim
detects this by analyzing the debuginfo function attributes for all
functions in a binary. Prior to LLVM 16, the flag which indicates
whether or not stack protection is enabled was not properly implemented
and so every function was marked as having stack protection enabled when
this was not actually the case.

Rust 1.70 updated to LLVM 16 and so Binskim can now detect the lack of
stack protection. This change enables stack protection for the
windows-msvc targets (and specifically disables it for uefi targets) to
resolve the Binskim issue.

Merged PR 26262: Disable stack protector for thumbv7a-*-windows-msvc targets

Disable stack protector for thumbv7a-*-windows-msvc targets

Enabling stack protection leads to LLVM asserting while trying to
generate relocations for some crates for this target. Since this target
is being deprecated in Windows anyway, just disable stack protection for
now.