Enable frame pointers for non-leaf functions on aarch64-linux
Without frame pointers, debuggers (all of WinDbg, gdb and lldb) are
unable to correctly walk the stack. This is especially bad for aborting
panics as `__rust_start_panic` and
`panic_abort::__rust_start_panic::abort` do not maintain the caller's
link register and so all debuggers stop unwinding at
`__rust_start_panic` which is entirely unhelpful.

By default, clang sets this option as well for aarch64-linux targets so
this makes sense for us to do as well.

Measuring code size for a few applications before and after this change:

| Project         | Before (bytes) | After (bytes) | Increase |
| --------------- | -------------- | ------------- | -------- |
| hello_world     |         660520 |        660544 |   0.004% |
| ripgrep         |       26331360 |      26425592 |   0.358% |
| ruff            |       31462976 |      31862240 |   1.269% |
| underhill.bin   |       23098428 |      23106684 |   0.036% |