Remove support for AES instructions in miri
These instructions are emulated using the aes crate which is banned by
the Crypto Board. We can request an exception to allow use of this crate
but such an exception requires on-going re-approval periodically. Rather
than incur that expense now, we're going to defer this until it can be
shown that users need this feature which will help support the case for
an exception to the Crypto board.